requirements.txt
pymongo
python-dotenv
requests
schedule